python3-flask-traced/requirements.txt
flask
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-ext-jaeger
opentelemetry-ext-flask
opentelemetry-ext-requests
